+++ src/updates.rs
@@ -1,15 +1,14 @@
 use crate::{
     block::BlockGenerator,
-    config::{Config, FileConfig, ThemeVars},
+    config::{BlockConfig, Config, FileConfig},
     utils::expand_tilde,
 };
 use colored::Colorize;
-use regex::{Regex, RegexBuilder};
 use std::{fs, process::exit};
 
 pub fn run(theme_name: String, config: &Config) {
-    let theme = match config.themes.get(&theme_name) {
-        Some(t) => t,
+    let vars = match config.themes.get(&theme_name) {
+        Some(t) => t.clone(),
         None => {
             log::error!("Theme `{theme_name}` is not listed in configuration file.");
             println!(
@@ -19,64 +18,87 @@ pub fn run(theme_name: String, config: &Config) {
             exit(1);
         }
     };
-    let update = UpdatesGenerator::new(theme_name, theme.clone());
+
+    let block_gen = BlockGenerator::new(
+        theme_name,
+        &vars,
+        FileConfig::Single(BlockConfig::default()),
+    );
+    let mut update_gen = UpdatesGenerator::new(block_gen);
 
     for (_, conf) in &config.files {
-        if let Some(contents) = update.generate(conf) {
-            fs::write(expand_tilde(&conf.path), contents.as_bytes()).unwrap();
+        for block in conf.to_blocks() {
+            let results = update_gen.generate(&block);
+            print_err(results, &block);
         }
+    }
+}
+
+fn print_err(results: Result<String, UpdatesError>, block: &BlockConfig) {
+    let path = &block.path;
+    match results {
+        Ok(s) => fs::write(expand_tilde(&path), s.as_bytes()).unwrap(),
+        Err(e) => match e {
+            UpdatesError::InvalidBlock => {
+                println!("Hello");
+                log::error!(
+                    "Failed to find Themer block (tag: {}) inside {path}",
+                    block.tag.clone().unwrap_or("No tag".to_string())
+                )
+            }
+            UpdatesError::UnableToRead => log::error!("Failed to read file {path}"),
+        },
     }
 }
 
 pub struct UpdatesGenerator {
-    theme_name: String,
-    theme: ThemeVars,
+    pub block_generator: BlockGenerator,
+}
+
+#[derive(Debug)]
+pub enum UpdatesError {
+    UnableToRead,
+    InvalidBlock,
 }
 
 impl UpdatesGenerator {
-    pub fn new(theme_name: String, theme: ThemeVars) -> Self {
-        Self { theme_name, theme }
-    }
-
-    pub fn generate(&self, config: &FileConfig) -> Option<String> {
-        let contents = self.get_file_contents(config);
-        if let Err(_) = contents {
-            return None;
+    pub fn new(gen: BlockGenerator) -> Self {
+        Self {
+            block_generator: gen,
         }
-        let contents = contents.unwrap();
-
-        let blk_gen = BlockGenerator::new(&self.theme_name, &self.theme, config);
-        let mut new_block = blk_gen.generate();
-        // Replacing dollar sign to avoid Regex issues
-        new_block = blk_gen.wrap(&new_block).replace("$", "$$");
-
-        Some(Self::get_block_re(&config.comment)
-            .replacen(&contents, 1, new_block)
-            .to_string())
     }
 
-    fn get_file_contents(&self, conf: &FileConfig) -> Result<String, ()> {
-        let contents = match fs::read_to_string(expand_tilde(&conf.path)) {
-            Ok(f) => f,
-            Err(e) => {
-                log::error!("Error reading `{}`:\n{e}", conf.path);
-                return Err(());
+    pub fn read_file(&self, path: &String) -> Result<String, UpdatesError> {
+        match fs::read_to_string(expand_tilde(&path)) {
+            Ok(f) => Ok(f),
+            Err(_) => {
+                return Err(UpdatesError::UnableToRead);
             }
-        };
-
-        let themer_block_re = Self::get_block_re(&conf.comment);
-        if !themer_block_re.is_match(&contents) {
-            log::warn!("Failed to find THEMER block inside of `{}`. Skipping it.", conf.path);
-            return Err(());
         }
-
-        Ok(contents)
     }
 
-    pub fn get_block_re(comment: &String) -> Regex {
-        RegexBuilder::new(&format!("{0} THEMER\n.*{0} THEMER_END", comment))
-            .dot_matches_new_line(true)
-            .build()
-            .unwrap()
+    pub fn validate_block(&self, contents: &String) -> Result<(), UpdatesError> {
+        if !self.block_generator.get_re().is_match(contents) {
+            return Err(UpdatesError::InvalidBlock);
+        }
+
+        Ok(())
+    }
+
+    pub fn generate(&mut self, config: &BlockConfig) -> Result<String, UpdatesError> {
+        self.block_generator.config = config.clone();
+
+        let contents = self.read_file(&config.path)?;
+        self.validate_block(&contents)?;
+
+        let mut update = self.block_generator.generate();
+        // Replacing dollar sign to avoid Regex issues
+        update = self.block_generator.wrap(&update).replace("$", "$$");
+
+        Ok(self
+            .block_generator
+            .get_re()
+            .replacen(&contents, 1, update)
+            .to_string())
     }
 }